packages/pages/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.7,<0.9.0"]
build-backend = "uv_build"

[project]
name = "lineo-pages"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "J. David Ibáñez", email = "jdavid.ibp@gmail.com" }
]
license = { text = "GPL-3.0-only" }
requires-python = ">=3.12"
dependencies = [
    "django-polymorphic",
]
classifiers = [
    "license :: osi approved :: gnu general public license v3 (gplv3)",
]

[tool.uv.build-backend]
module-root = "."
